Major update
Graph implementation: added abstract class to implement graphs with
adjacency lists and undirected graphs. Implemented unweighted and
weighted graphs.

Added new functions to compute distances and paths in graphs that are
not assumed to be unweighted. The old functions doing this work assume
the graphs to be so.

Input graphs: renamed existing files and added new inputs (the same
graphs but with weights on edges).

Reorganised the code to debug the functions of path- and
distance-finding.

Classes that implement graph paths are now templates.