Path-relative inclusion of headers
Each file containing the implementation of class or functions now use
the "path-relative way" of including files.

For example, the implementation of class node_path defined in header
node_path.hpp has its implementation in node_path.cpp. Since this is a
template, the file node_path.hpp has to include its implementation file
and it does so with
	<lgraph/data_structures/node_path.cpp>
The implementation file also includes the header (so that Qt can
highlight functions and attributes properly) similarly:
	<lgraph/data_structures/node_path.hpp>